WiP: Build java using GraalVM on the sfdk target
Reworks the compile script to build the java source on a local GraalVM
inside an sfdk target, rather than on a remote device.

The changes add restrictions to the amount of heap the JVM can allocate.
This is necessary to appease the QEMU version used by the sfdk build
engine.

The memory restrictions unfortunately prevent the build from going
through, so this is a proof-of-concept rather than a working solution.